changed buf size, added custom input function, added partition tool confirmation
The size of buf was 256, but the valid inputs were "1" and "2" , and other invalid inputs with n length. So I added a custom input function that uses getchar to fill the character array, and only scans len characters, which in this case is 1. So it will only scan the first character of input. I also added a confirmation, So that if the user enters something by mistake, they can just type N, and then enter the correct choice.